+++ src/contexts/GlobalContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+
+export const GlobalContext = createContext()
+
