+++ src/pages/PokedexPage/PokedexPage.js
@@ -2,9 +2,14 @@ import Card from "../../components/Card/Card";
 import { Container } from "./PokedexPage.styled";
 import Header from "../../components/Header/Header";
 import { BASE_URL } from "../../constants/url";
+import { useContext } from "react";
+import { GlobalContext } from "../../contexts/GlobalContext";
 
-function PokedexPage(props) {
-  const { pokedex, removeFromPokedex } = props;
+function PokedexPage() {
+
+  const context = useContext(GlobalContext)
+
+  const { pokedex, removeFromPokedex } = context;
 
   return (
     <Container>
